chore: remove beholder
As part of the TB 2.3.0 release, TensorBoard has announced that we would
discontinue support for beholder.

We would definitely welcome any contribution to make it work with TF 2.x
and the plugin as a dynamic plugin.
