+++ tensorboard/components/tf_tensorboard/default-plugins.html
@@ -44,6 +44,5 @@ limitations under the License.
   rel="import"
   href="../tf-profile-redirect-dashboard/tf-profile-redirect-dashboard.html"
 />
-<link rel="import" href="../tf-beholder-dashboard/tf-beholder-dashboard.html" />
 <link rel="import" href="../tf-hparams-dashboard/tf-hparams-dashboard.html" />
 <link rel="import" href="../tf-mesh-dashboard/tf-mesh-dashboard.html" />
