+++ tensorboard/pip_package/BUILD
@@ -40,7 +40,6 @@ sh_binary(
         "//tensorboard:lib",  # User-facing overall TensorBoard API
         "//tensorboard:version",  # Version module (read by setup.py)
         "//tensorboard/data/experimental:experiment_from_dev",
-        "//tensorboard/plugins/beholder",  # User-facing beholder API
         "//tensorboard/plugins/hparams",  # User-facing hparams API
         "//tensorboard/plugins/mesh",  #  User-facing mesh API
         "//tensorboard/plugins/projector",  # User-facing projector API
