+++ tensorboard/pip_package/test_pip_package.sh
@@ -156,12 +156,11 @@ from tensorboard.plugins.hparams import summary_v2 as hp
 hp.hparams_pb({'optimizer': 'adam', 'learning_rate': 0.02})
 "
   if [ -n "${tf_version}" ]; then
-    # Only test summary scalar, beholder, and mesh summary
+    # Only test summary scalar and mesh summary
     python -c "
 import tensorboard as tb
 tb.summary.v1.scalar_pb('test', 42)
 tb.summary.scalar('test v2', 1337)
-from tensorboard.plugins.beholder import Beholder, BeholderHook
 from tensorboard.plugins.mesh import summary
 "
   fi
