--- tensorboard/plugins/beholder/__init__.py
@@ -1,23 +1,0 @@
-# Copyright 2017 The TensorFlow Authors. All Rights Reserved.
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-# Only import Beholder API when tensorflow is available.
-try:
-    # pylint: disable=unused-import
-    import tensorflow
-except ImportError:
-    pass
-else:
-    from tensorboard.plugins.beholder.beholder import Beholder
-    from tensorboard.plugins.beholder.beholder import BeholderHook
