--- tensorboard/plugins/beholder/tf_beholder_dashboard/BUILD
@@ -1,36 +1,0 @@
-load("//tensorboard/defs:web.bzl", "tf_web_library")
-
-package(default_visibility = ["//tensorboard:internal"])
-
-licenses(["notice"])  # Apache 2.0
-
-tf_web_library(
-    name = "tf_beholder_dashboard",
-    srcs = [
-        "tf-beholder-dashboard.html",
-        "tf-beholder-info.html",
-        "tf-beholder-video.html",
-    ],
-    path = "/tf-beholder-dashboard",
-    deps = [
-        "//tensorboard/components/tf_backend",
-        "//tensorboard/components/tf_card_heading",
-        "//tensorboard/components/tf_categorization_utils",
-        "//tensorboard/components/tf_color_scale",
-        "//tensorboard/components/tf_dashboard_common",
-        "//tensorboard/components/tf_imports:d3",
-        "//tensorboard/components/tf_imports:lodash",
-        "//tensorboard/components/tf_imports:polymer",
-        "//tensorboard/components/tf_runs_selector",
-        "//tensorboard/components/tf_tensorboard:plugin_dialog",
-        "//tensorboard/components/tf_tensorboard:registry",
-        "@org_polymer_paper_button",
-        "@org_polymer_paper_dialog",
-        "@org_polymer_paper_icon_button",
-        "@org_polymer_paper_listbox",
-        "@org_polymer_paper_radio_group",
-        "@org_polymer_paper_slider",
-        "@org_polymer_paper_spinner",
-        "@org_polymer_paper_tooltip",
-    ],
-)
