--- tensorboard/plugins/beholder/tf_beholder_dashboard/tf-beholder-info.html
@@ -1,139 +1,0 @@
-<!--
-@license
-Copyright 2016 The TensorFlow Authors. All Rights Reserved.
-
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-    http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
--->
-
-<link rel="import" href="../tf-imports/polymer.html" />
-<link rel="import" href="../tf-backend/tf-backend.html" />
-
-<dom-module id="tf-beholder-info">
-  <template>
-    <template is="dom-repeat" items="[[_items]]">
-      <div class="section-info" style$="height: [[item.height]]px">
-        <ul>
-          <li>[[item.name]]</li>
-          <li>shape: [[item.shape]]</li>
-          <li>range: [ [[item.min]], [[item.max]] ]</li>
-          <li>mean: [[item.mean]]</li>
-        </ul>
-      </div>
-    </template>
-
-    <style>
-      .section-info {
-        margin: 0 0 5px 0;
-      }
-      .section-info ul {
-        list-style-type: none;
-        margin: 0;
-        padding-left: 10px;
-      }
-    </style>
-  </template>
-
-  <script>
-    'use strict';
-    (function() {
-      const INFO_ROUTE = tf_backend
-        .getRouter()
-        .pluginRoute('beholder', '/section-info');
-
-      Polymer({
-        is: 'tf-beholder-info',
-        properties: {
-          fps: {
-            // Not actually FPS!
-            type: Number,
-            value: 10,
-            observer: '_fpsChanged',
-          },
-
-          xhrTimeout: {
-            type: Number,
-            value: 10000,
-          },
-
-          _items: {
-            type: Array,
-            value: () => [{name: 'Loading...'}],
-          },
-
-          _xhr: Object,
-          _timer: Number,
-        },
-
-        attached() {
-          this._load();
-        },
-
-        detached() {
-          this._clear();
-        },
-
-        _load() {
-          this._clear();
-          this._xhr = new XMLHttpRequest();
-          this._xhr.open('GET', INFO_ROUTE, true);
-          this._xhr.timeout = this.xhrTimeout;
-          this._xhr.onload = this._onLoad.bind(this);
-          this._xhr.onerror = this._retry.bind(this, this._getSleep());
-          this._xhr.ontimeout = this._retry.bind(this, 1);
-          this._xhr.send(null);
-        },
-
-        _onLoad() {
-          if (this._xhr.status == 200) {
-            const response = JSON.parse(this._xhr.responseText);
-            console.assert(
-              Array.isArray(response),
-              'Expected response to be in an array'
-            );
-            this._items = response;
-          }
-          this._retry(this._getSleep());
-        },
-
-        _retry(delay) {
-          this._timer = window.setTimeout(this._load.bind(this), delay);
-        },
-
-        _getSleep() {
-          return 1000 / (this.fps === 0 ? 1 : this.fps);
-        },
-
-        _clear() {
-          if (this._timer) {
-            window.clearTimeout(this._timer);
-            this._timer = null;
-          }
-          if (this._xhr) {
-            if (this._xhr.readyState < XMLHttpRequest.DONE) {
-              this._xhr.abort();
-            }
-            this._xhr = null;
-          }
-        },
-
-        _fpsChanged(newValue, oldValue) {
-          if (newValue == 0) {
-            this._clear();
-          } else if (oldValue == 0) {
-            this._load();
-          }
-        },
-      });
-    })();
-  </script>
-</dom-module>
