--- tensorboard/plugins/beholder/tf_beholder_dashboard/tf-beholder-video.html
@@ -1,138 +1,0 @@
-<!--
-@license
-Copyright 2016 The TensorFlow Authors. All Rights Reserved.
-
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-    http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
--->
-
-<link rel="import" href="../tf-imports/polymer.html" />
-<link rel="import" href="../tf-backend/tf-backend.html" />
-
-<dom-module id="tf-beholder-video">
-  <template>
-    <div id="container">
-      <img id="video" src$="[[_imageURL]]" />
-    </div>
-
-    <style>
-      img {
-        image-rendering: pixelated;
-        margin-right: 10px;
-      }
-    </style>
-  </template>
-
-  <script>
-    'use strict';
-    (function() {
-      const BLANK =
-        'data:image/gif;base64,R0lGODlhAQABAAD/ACwAAAAAAQABAAACADs=';
-      const FEED_ROUTE = tf_backend
-        .getRouter()
-        .pluginRoute('beholder', '/beholder-frame');
-      const PING_ROUTE = tf_backend
-        .getRouter()
-        .pluginRoute('beholder', '/ping');
-
-      Polymer({
-        is: 'tf-beholder-video',
-        properties: {
-          // Only used for determining ping frequency.
-          fps: {
-            type: Number,
-            value: 10,
-            observer: '_fpsChanged',
-          },
-
-          pingSleep: {
-            type: Number,
-            value: 1000,
-          },
-
-          xhrTimeout: {
-            type: Number,
-            value: 2500,
-          },
-
-          _imageURL: {
-            type: String,
-            value: BLANK,
-          },
-
-          _xhr: Object,
-          _timer: Number,
-          _isDead: Boolean,
-        },
-
-        attached() {
-          this.set('_imageURL', FEED_ROUTE);
-          this._ping();
-        },
-
-        detached() {
-          this._clear();
-          this.set('_imageURL', BLANK);
-        },
-
-        _ping() {
-          this._clear();
-          this._xhr = new XMLHttpRequest();
-          this._xhr.open('GET', PING_ROUTE, true);
-          this._xhr.timeout = this.xhrTimeout;
-          this._xhr.onload = this._onPingLoad.bind(this);
-          this._xhr.onerror = this._onPing.bind(this, false, this.pingSleep);
-          this._xhr.ontimeout = this._onPing.bind(this, false, 1);
-          this._xhr.send(null);
-        },
-
-        _onPingLoad() {
-          if (this._xhr.status == 200) {
-            const response = JSON.parse(this._xhr.responseText);
-            this._onPing(response['status'] == 'alive', this.pingSleep);
-            return;
-          }
-          this._onPing(false, this.pingSleep);
-        },
-
-        _onPing(isAlive, retryMs) {
-          if (isAlive && this._isDead) {
-            this.set('_imageURL', FEED_ROUTE + '?t=' + new Date().getTime());
-          }
-          this._isDead = !isAlive;
-          this._timer = window.setTimeout(() => this._ping(), retryMs);
-        },
-
-        _clear() {
-          if (this._timer) {
-            window.clearTimeout(this._timer);
-            this._timer = null;
-          }
-          if (this._xhr) {
-            if (this._xhr.readyState < XMLHttpRequest.DONE) {
-              this._xhr.abort();
-            }
-            this._xhr = null;
-          }
-        },
-
-        _fpsChanged(newValue, oldValue) {
-          if (newValue == 0) {
-            this._clear();
-          } else if (oldValue == 0) {
-            this._ping();
-          }
-        },
-      });
-    })();
-  </script>
-</dom-module>
